+++ src/mobx/stores/bridgeStore.ts
@@ -30,6 +30,7 @@ import { RenVMTransaction } from '../model/bridge/renVMTransaction';
 import { defaultNetwork } from 'config/networks.config';
 import { NetworkStore } from './NetworkStore';
 import { Network } from 'mobx/model/network/network';
+import { REN_FEES_ENDPOINT } from '../../config/constants';
 
 export enum Status {
 	// Idle means we are ready to begin a new tx.
@@ -135,7 +136,7 @@ class BridgeStore {
 		// NB: At construction time, the value of wallet provider is unset so we cannot fetch network
 		// from provider. Align network init logic w/ how it works in the walletStore.
 		const network = defaultNetwork;
-		this.network = network ? network.name : '';
+		this.network = network ? network.symbol : '';
 
 		extendObservable(this, {
 			...defaultProps,
@@ -196,7 +197,7 @@ class BridgeStore {
 				// Set shortened addr.
 				const { network } = this.store.network;
 				// NB: Only ETH supported for now.
-				if (network.name !== NETWORK_LIST.ETH) return;
+				if (network.symbol !== NETWORK_LIST.ETH) return;
 				await this.reload();
 			},
 		);
@@ -271,7 +272,8 @@ class BridgeStore {
 		this.shortAddr = shortenAddress(connectedAddress);
 
 		this.loading = true;
-		Promise.all([
+
+		return Promise.all([
 			// Fetch old transactions and reload any incomplete tx.
 			this._fetchTx(connectedAddress),
 			this._getBalances(connectedAddress),
@@ -553,23 +555,47 @@ class BridgeStore {
 		const { queueNotification } = this.store.uiState;
 		try {
 			await retry(async () => {
-				const [badgerBurnFee, badgerMintFee, renvmBurnFee, renvmMintFee] = (
+				const [badgerBurnFee, badgerMintFee] = (
 					await Promise.all([
 						this.adapter.methods.burnFeeBps().call(),
 						this.adapter.methods.mintFeeBps().call(),
-						this.gateway.methods.burnFee().call(),
-						this.gateway.methods.mintFee().call(),
 					])
 				).map((result: number) => result / MAX_BPS);
 				this.badgerMintFee = badgerMintFee;
 				this.badgerBurnFee = badgerBurnFee;
-				this.renvmBurnFee = renvmBurnFee;
-				this.renvmMintFee = renvmMintFee;
 			}, defaultRetryOptions);
 		} catch (err) {
 			console.error(err);
 			queueNotification(`Failed to fetch fees: ${err.message}`, 'error');
 		}
+	};
+
+	_saveRenFees = (burnFee: number, mintFee: number): void => {
+		this.renvmMintFee = mintFee;
+		this.renvmBurnFee = burnFee;
+	};
+
+	_getRenFees = async (): Promise<void> => {
+		const { queueNotification } = this.store.uiState;
+		fetch(REN_FEES_ENDPOINT, {
+			method: 'POST',
+			headers: {
+				'Content-Type': 'application/json',
+				Accept: 'application/json',
+			},
+			body: JSON.stringify({ method: 'ren_queryBlockState', id: 1, jsonrpc: '2.0', params: {} }),
+		})
+			.then((res: { json: () => Promise<Response> }) => res.json())
+			.then((json: any) =>
+				this._saveRenFees(
+					json.result.state.v.BTC.fees.chains[3].burnFee / MAX_BPS,
+					json.result.state.v.BTC.fees.chains[3].mintFee / MAX_BPS,
+				),
+			)
+			.catch((err: Error) => {
+				queueNotification(`Failed to fetch RenVM Fees: ${err.message}`, 'error');
+				console.error(err);
+			});
 	};
 
 	_getBalances = async (userAddr: string): Promise<void> => {